forgor which change was mine lol

+++ include/m_field_info.h
@@ -37,7 +37,7 @@ typedef enum FieldRoom{
 // void func_80087C30_jp();
 // void func_80087C40_jp();
 // void func_80087C64_jp();
-s32 mFI_GetFieldId(void);
+u16 mFI_GetFieldId(void);
 // void func_80087C9C_jp();
 // void func_80087D30_jp();
 // void func_80087DC8_jp();
